+++ packages/table-react/src/Table.tsx
@@ -1,20 +1,73 @@
-import React from "react";
-import { TableAccordionRowData, TableAnchorRowData, TableRow } from "./TableRow";
+import React, { ReactNode } from "react";
+import { TableRowAnchor } from "./TableRowAnchor";
+import { TableRowAccordion } from "./TableRowAccordion";
+import { TableRowSimple } from "./TableRowSimple";
 
-export type TableRowData = string[] | TableAnchorRowData | TableAccordionRowData;
+//-------TODO:  Skulle vi ha flyttet alle disse interfacene, typene (og type guardsene?) ut i en egen fil? -------------
+export type TableRowData = ReactNode[] | TableRowAnchorData | TableRowAccordionData; // Endret fra string. Greit?
+
+export type TableRowDataVariants = "anchor" | "accordion";
+
+export interface BaseTableRowData {
+    type: TableRowDataVariants;
+}
+
+export interface TableRowAnchorData extends BaseTableRowData {
+    rowData: string[];
+    type: "anchor";
+    href: string;
+    hrefLabel: string;
+    onRowClick?: (href: string) => void;
+}
+
+export interface TableRowAccordionData extends BaseTableRowData {
+    rowData: string[];
+    type: "accordion";
+    elementLabel: string;
+    defaultOpen: boolean;
+    children: ReactNode[];
+}
+
+export function isAnchorRowData(
+    row: ReactNode[] | TableRowAnchorData | TableRowAccordionData,
+): row is TableRowAnchorData {
+    // Endret fra string. Greit?
+    return "type" in row && row.type === "anchor";
+}
+
+export function isAccordionRowData(
+    row: ReactNode[] | TableRowAnchorData | TableRowAccordionData,
+): row is TableRowAccordionData {
+    // Endret fra string. Greit?
+    return "type" in row && row.type === "accordion";
+}
+
+// ---------------------------------------------------------------------------------------------------------------------
 
 interface Props {
     columns: string[];
     rows: Array<TableRowData>;
     className?: string;
-    accordiation?: boolean;
+    accordion?: boolean;
 }
 
-export function Table({ columns, rows, className, accordiation }: Props) {
+export function Table({ columns, rows, className, accordion }: Props) {
     const componentClassName = "jkl-table".concat(
-        accordiation ? ` jkl-table--accordiation` : "",
+        accordion ? ` jkl-table--accordion` : "",
         className ? ` ${className}` : "",
     );
+
+    const createAppropriateTableRow = (row: TableRowData, index: number) => {
+        // TODO: Good/bad practice å ha keys her?
+
+        if (isAnchorRowData(row)) {
+            return <TableRowAnchor key={index} row={row} />;
+        } else if (isAccordionRowData(row)) {
+            return <TableRowAccordion key={index} row={row} />;
+        } else {
+            return <TableRowSimple key={index} row={row} />;
+        }
+    };
 
     return (
         <table className={componentClassName}>
@@ -22,7 +75,8 @@ export function Table({ columns, rows, className, accordiation }: Props) {
                 {columns.map((columnValue, i) => (
                     <col key={i} />
                 ))}
-                <col className="jkl-table__column--narrow" />
+                {/* Icon-column has fixed width */}
+                <col className="jkl-table__column--icon" />
             </colgroup>
             <thead>
                 <tr className="jkl-table__row">
@@ -33,11 +87,7 @@ export function Table({ columns, rows, className, accordiation }: Props) {
                     ))}
                 </tr>
             </thead>
-            <tbody>
-                {rows.map((row, i) => (
-                    <TableRow row={row} key={i} />
-                ))}
-            </tbody>
+            <tbody>{rows.map((row, i) => createAppropriateTableRow(row, i))}</tbody>
         </table>
     );
 }