+++ packages/table-react/src/TableRowAnchor.tsx
@@ -1,12 +1,13 @@
-import { TableAnchorRowData } from "./TableRow";
-import { TableData } from "./TableData";
 import React from "react";
+import { TableRowAnchorData } from "./Table";
+import { TableData } from "./TableData";
 
 interface PropTypes {
-    row: TableAnchorRowData;
+    row: TableRowAnchorData;
 }
 
-export function AnchorRow({ row }: PropTypes) {
+export function TableRowAnchor({ row }: PropTypes) {
+
     const onClick = () => {
         if (row.onRowClick) {
             row.onRowClick(row.href);
@@ -18,8 +19,10 @@ export function AnchorRow({ row }: PropTypes) {
     return (
         <tr onClick={onClick} className={"jkl-table__row jkl-table__row--data-row jkl-table__row--clickable-row"}>
             {row.rowData.map((data, j) => (
-                <TableData data={data} row={row} key={j} isFirstCell={j === 0} />
+                <TableData key={j} data={data} row={row} isFirstCell={j === 0} />
             ))}
+            {/* Empty last cell in stead of icon (see TableRowAccordion) */}
+            {/* This is necessary for 'hover' function to work properly  */}
             <td />
         </tr>
     );